Add article skeleton and hotel details
There are a couple of bugs in the way the hotel details are displaying.

The total price not being limited to 2 decimal places is fixed in a different branch, so I'll ignore that.

Dates that should read '20th' are being displayed as '20undefined'. Not sure why, but I'd rather fix in a separate branch than have coding fixes in article branches.