Use `<img>` instead of Next `<Image>`
Perhaps this is not ideal and only a temporary fix, but it does fix the bug at hand: Using the `<img>` tag there is no flicker.

On previous projects I've often found Next's `<Image>` tag to not only be buggy, but to quickly use resources when hosting, so I'm happy with this as at least a temporary solution.