Table more or less styled as wanted
Only minor issue, is it needs a height in order to sticky the top row. I have this set to 75vh. `max-height` seems to make no difference, so on bigger screens, the content doesn't quite fill the table.

Almost OK on a 13" screen, so will just need to visit some more places so it fills on all screen sizes.