Add link to the article path (if one exists)
Only working with dummy path for now; haven't figured-out the exact path.

This hijacks the `<Card.Link>` component without using the rest of the card, which is probably bad practice, but I figured better than rewriting the same code.

Not entirely sure I'm a fan of linking the entire card to the article like this. If all criteria have a description, then it'll make it a pretty big link, especially on mobile.

Perhaps each individual description should be a link to the article instead.