Rename article (and folder)
Note that nothing I tried could get this to run with `npm run dev`. It would still build, but wouldn't run in dev mode. I even tried restarting my laptop.

Eventually I went nuclear and delete the `next` folder (wasn't sure of the consequences of that) and that fixed it, so remember that for next time.