Add `GBP` suffix to accommodation paid for in £'s
I think it's a bit clearer to establish the currency. For someone who's never been to Sofia for example, they might not know that `£` is pounds; they might think it represents the local currency, because tbf it is a bit confusing that I'm paying pounds in countries that don't use it.