Give countries a minimum height and move stay details to right on mobile
Using `Math.max` and including `96` for the country `minHeight` means that no matter the length of the stay, the element will always be the same height as the title so the display won't be compromised.

The other changes move all stay details to the right side of the timeline on mobile. This prevents that title from scrolling over the stay details.