Fix arrival date matching timeline end date
Updating so the comparison is `Date.parse(stay[dates][arrival]) < Date.parse(endDate)` means that if the timeline end date is the same as a places arrival date, then it'll be omitted altogether.

As the timeline is my nights spent somewhere, at that point I'd have spent no nights there, so I think better to remove it altogether.