Move `after` pseudo-element to separate element
The bug in the last commit was using an invalid Tailwind class (`border-zinc-10` instead of `border-zinc-100`).

In the process of fixing this, I moved the `after` pseudo-element to a separate `div` because I thought that the bug was that `dark:after` wouldn't work.

While this didn't fix the actual bug, I find it much more readable, and according to the Tailwind docs `It’s worth noting that you don’t really need ::before and ::after pseudo-elements for most things in Tailwind projects — it’s usually simpler to just use a real HTML element`, so I'm going to keep it.