Update stay max height to be `px` rather than `vh` value
Using the viewport height means that for timelines with multiple stays that exceed `150vh` in size, whenever the viewport resizes (for example, on mobile when scrolling down and the menu bar hides) there is a 'jerk' as the stays resize to accommodate the new viewport height. This is especially true when at the bottom of the page and multiple elements above are resized.

Using an absolute size isn't ideal but `1050px` feels a nice middle-ground to look alright on desktop and mobile.