Fix production error
Was getting an error that wasn't present in the development that I think was caused by the `<Link>` tag not having a `href` value.

Presumably the `Component.render && Component.render.name === 'LinkComponent'` check does not immediately return `true`.

While this solution is a bit undesirable, as it would make reusing this component as anything other than a `div` or a `Link` a bit of a pain, I don't envision using it elsewhere anytime soon, and more likely it'll be simplified when deleting the 'Socials' page, so why make things more complicated than they need to be?