Return departure date rather than todays date if departure date is later than today
The bug was that when fetching the timeline data in `getStaticProps`, returning the departure date as today's date only works when a build was done that day.

If a build is done one day, and you view the timeline the next day, the departure date is then out of date. so instead returning the departure date and then parsing this date in the timeline component, the displayed departure date will always be correct.

Note that one related bug still exists:

While this will correctly display stays in progress at build time, what about a future stay? For example, if I build the app on the 1st of the month, and have a new stay starting on the 2nd, that stay won't be returned at all because it was out of the date range at build-time.