Return future stays to timeline
This commit fixes the bug mentioned in the previous commit, where future stays wouldn't be included in the timeline array at build time so if, for example, I build the app on the 1st, and have a stay that starts on the 2nd, then even after the 2nd, this stay would not show until the app rebuilds.

To fix this, for the 'current' (the Digital nomad) timeline, we pass-in and end date of `null` (rather than `undefined`).

In the `parsedTimelineData` function, we then set the `endDateIsSet` variable to `false` if the end date is null, before we set the end date to today's date.

When parsing the timeline data, we then include any future stays if `endDateIsSet` is set to false. This means that if today is the 1st, a stay that starts on the 2nd will be included in the data set at build time.

Obviously we do not want to display these future stays until they are in the present, so in the `TimelineComponent` we only display stays if `Date.parse(stay[arrival]) < beginningOfToday`. We then parse the end date to be today, and build the stay based off that date.

It would be feasible that countries with only future stays would be included in the timeline data, at which point these countries would be included in the timeline, but would be empty. To mitigate this, we parse the timeline data to check that past or present stays exist, and remove any countries where this is not true.

This gives us the best of both worlds, where we can still use `getStaticProps` to parse (most) of the timeline data at build time, but client-side we read this data differently depending on the date.

Works as intended (I think), but still needs a small refactor.