Add compact mode height transition
You cannot add a transition to/from `height: auto`, hence adding `height: '106px'` for when in compact mode.

This transition transcends timelines, where switching from, for example, the digital nomad timeline to the backpacking timeline, the elements of the new timeline transition from the old timeline.

This is the reason for setting `stayOrderFirst` to always be `true` for the first element of a timeline. When the first stay being on the left/right differs between timelines, it's very jerky when the timeline elements are transitioning.

I prefer it when they all start in the same place anyway.