Set stay `min-height` instead of `height`
Setting the `height` to 106px meant that when a stay wrapped onto three lines (which happens on 375px phone screens for stays like `Bali`) the stay didn't expand and the content overlapped the element padding, which looked a bit shit.

Setting the `min-height` instead means that in this case the element will expand when its content wraps to three lines.

The `Math.min()` function is because `min-height` takes precedence over `max-height` to really long stays were expanding beyond 1050px. That can get a little out of hand and be multiple times the viewport height, hence why I added the `max-height` in the first place.

Remove the `lg` from `lg:flex` means that the stay element becomes a flex-element and grows to fit the container.