Update the stay `min-height` yet again
For short stays, it was actually bigger for compact mode than detailed mode because 106px was bigger than the calculated height.

Instead returning the minimum of 1050, and the maximum of 106 and the calculated height means that the minimum will always be 106px, saving this annoying height change for short stays.