Attempt to consider user's timezone when generating timeline
Fuck I hate timezones.

This commit reverts the previous change of adding a timezone to the `formatDate` function, which wasn't working anyway because I'd used `timezone` instead of `timeZone`, because that's the kind of genius I am, but I also decided this wasn't the place to mess around with timezones and any date passed-in here shoiuld be formatted in UTC anyway so that it will be rendered the same no matter where the user is (so long as it's passed-in as a UTC date).

Instead, the key change is that we're using the local date for the user in the timeline component (imported into the timeline in the `usersDate` variable) and comparing that date to the arrival/departure in place of the `beginningOfToday` variable, which was the beginning of the current day in UTC time, but was inaccurate if the date for the current user differed for the current UTC date.

This, in theory, will keep the timeline accurate for the local user at all hours of the day so even after midnight, my current stay should say 'now' instead of a date.

That was a lot of work just to say 'now' for a bit longer.

As the `formatDate` function is used in a lot of places, all should be checked (before and after midnight) as displaying dates correctly, particularly the timeline.