Compare dates rather than date strings
So it turns out that comparing date strings, for example `'2023-11-01' > '2023-11-02'` compares them lexicographically rather than chronologically, because fuck me that's why. So in order to make the comparisons I need, it's necessary to convert them to dates (using `new Date()`) first.

`new Date()` seems more versatile than `Date.parse()` so I've updated all instances of dates to use `new Date()` in this commit.

Worth noting that in the `dateStringIsToday` function, where we use an equality comparison (`===`) I've added `getTime()` to each of the dates. This is because when comparing date objects, `===` checks whether the two objects are the same in memory, not whether they have the same value, so `new Date('2023-11-01') === new Date('2023-11-01')` returns `false`. Adding `getTime()` to each date returns a numeric value and comparing numbers with `===` actually works.

I've also udpated the timeline end date to be `usersDate` rather than `new Date().toJSON().slice(0, 10)`, meaning that the timeline end date will adjust to the user's timezone.

Now I need to stay up until midnight again to see if it still works. I guess I'll go to the pub.