Remove future stays from timeline client-side
In an earlier iteration of the timeline, future stays were removed server-side in via `getStaticProps`, but as this only ran at build time, these future stays would get added back in as time progressed until the app was rebuilt, so I left them in the data and just hid them from the timeline.

A bug I found yesterday was that just hiding these stays caused a slight misconfiguration of the timeline styling, as the border at the top/bottom (depending whether viewing it as ascending or descending) was accounting for these hidden stays and displaying where it shouldn't.

This update removes any future stays from the timeline, in the timeline component but before the return (so from a styling perspective, these future stays don't actually exist).

This meant that we could remove the `if (new Date(stay[arrival]) < new Date(usersDate))` line from the timeline itself.

Got too drunk and forgot to check that `now` was being displayed after midnight last night, so I'll have to do it tonight instead if my hangover's gone by then.