First pass at including all images in the page source
This works in the sense that all images are included in the page source yet only one images is visible at a time and you can scroll between them.

However, given that it sort of starts with a pile of images (with only the top one visible) and scrapes the top on off each time you scroll right, then makes the next image visible, going in the opposite direction, the transitions don't work as wanted because the image has to come in first.