Rewrite click/swipe functionality
This commit rewrites the click and swipe functionality to scroll through an image stack. It adds swiping left and right, and ensures that swiping doesn't 'double up' (register as a swipe and as a click) by only running each if `imageIndexRef.current` is not null.

This ensures that the previous enter/exit has finished before the next swipe/click is permitted.