Update touch action functionality
This commit adds tracking of image touches on the Y axis. The main purpose of this is to allow the page to scroll as normal when the user scrolls more vertically than horizontally on an image, but to prevent page scrolling and instead scroll the image stack when the user scrolls more horizontally than vertically.

In place of the `event.preventDefault()` in `handleImageTouchStart`, which is what previously prevented the page from scrolling when a user swiped on an image (no matter what direction they swiped), I've added the `handleImageTouchMove` function. This function checks whether scrolling on the X axis or the Y axis is greater. If it is greater on the X axis (by more than 10 to allow the user a bit of leeway in order to establish their gesture before acting), then `event.preventDefault()` is run to stop the page from scrolling. If it is greater on the Y axis (by more than 10), then the starting X value state is set to null, meaning `handleWindowTouchEnd` has no impact and the image stack doesn't scroll.