Add check for pinching
This commit adds a check for the number of fingers that the user is touching the image with. In the event that it's more than one (for pinching), then it sets `setImageTouchStartX(null)` in order to prevent the image stack from scrolling. This will allow the user to zoom in on images.

Might want to update this check to be on the window rather than on the image, as it's conceivable, particularly on mobile, that one of the users fingers will be inside the image, and one outside, meaning that this check will not work.