Add `handleWindowTouchMove` function
Hard to accurately test this on the simulator so pushing to the review app to test. The idea is that if a user touches the screen with 2+ fingers (or anything else) then `imageTouchStartX` (as well as `imageTouchStartY`) is set to `null`, which will prevent the image stack from scrolling.

This should mean that the user can zoom into the images on a touch screen without the images scrolling whenever they remove their fingers.